+++ src/libexecline/el_parse_from_buffer.c
@@ -1,12 +1,13 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/buffer.h>
 #include <skalibs/stralloc.h>
 #include <execline/execline.h>
 
 static int next (unsigned char *c, void *p)
 {
-  register int r = buffer_get((buffer *)p, (char *)c, 1) ;
+  register ssize_t r = buffer_get((buffer *)p, (char *)c, 1) ;
   if (r < 0) return 0 ;
   if (!r) *c = 0 ;
   return 1 ;
